requirements.txt
Django==3.1.7
ipython==7.22.0
psycopg2-binary
djangorestframework==3.12.4
python-dotenv==0.17.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
gunicorn==20.1.0